server/pyproject.toml
[tool.poetry]
name = "Backend stub"
version = "1.5.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.2"
psycopg2-binary = "2.8.6"
gunicorn = "20.1.0"
ipython = "7.22.0"
djangorestframework = "3.12.4"
dj-database-url = "^0.5.0"
djangorestframework-simplejwt = "^5.2.0"
django-filter = "^22.1"
celery = "^5.2.7"
redis = "^4.3.4"
openpyxl = "^3.0.10"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "3.2"
pep8-naming = "0.11.1"
flake8 = "3.9.0"
flake8-import-order = "0.18.1"
flake8-blind-except = "0.2.0"
flake8-builtins = "1.5.3"
flake8-broken-line = "0.3.0"
flake8-commas = "2.0.0"
flake8-eradicate = "1.0.0"
flake8-print = "4.0.0"
flake8-quotes = "3.2.0"
flake8-cognitive-complexity = "0.1.0"
flake8-bugbear = "21.4.3"
isort = "5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
